relocate shared functions
