+++ apps/playwright/fixtures/AppPage.ts
@@ -1,7 +1,18 @@
 import { type Page, expect } from "@playwright/test";
 
-import { Test_MeQueryDocument } from "~/utils/user";
+import { CreateUserInput, User } from "@gc-digital-talent/graphql";
+
+import {
+  defaultUser,
+  Test_CreateUserMutationDocument,
+  Test_MeQueryDocument,
+} from "~/utils/user";
 import { FeatureFlags, getFeatureFlagConfig } from "~/utils/featureFlags";
+import { GraphQLResponse } from "~/utils/graphql";
+import {
+  Test_RolesQueryDocument,
+  Test_UpdateUserRolesMutationDocument,
+} from "~/utils/roles";
 
 import { getAuthTokens } from "../utils/auth";
 
@@ -87,6 +98,33 @@ class AppPage {
     const res = await this.graphqlRequest(Test_MeQueryDocument);
 
     return res.me;
+  }
+
+  async createUser(user?: Partial<CreateUserInput>): Promise<User> {
+    return this.graphqlRequest(Test_CreateUserMutationDocument, {
+      user: {
+        ...defaultUser,
+        ...user,
+      },
+    }).then((res: GraphQLResponse<"createUser", User>) => res.createUser);
+  }
+
+  async addRolesToUser(userId: string, roles: string[], team?: string) {
+    const allRoles = await this.graphqlRequest(Test_RolesQueryDocument);
+    const roleInputArray = allRoles.roles
+      .filter((role) => roles.includes(role.name))
+      .map((role) => {
+        return { roleId: role.id, teamId: team };
+      });
+
+    await this.graphqlRequest(Test_UpdateUserRolesMutationDocument, {
+      updateUserRolesInput: {
+        userId,
+        roleAssignmentsInput: {
+          attach: roleInputArray,
+        },
+      },
+    });
   }
 }
